# ddl: reduce test time when running tests with `race`