--- a/ddl/syncer.go
+++ b/ddl/syncer.go
@@ -38,14 +38,13 @@ const (
 	DDLGlobalSchemaVersion = "/tidb/ddl/global_schema_version"
 	// InitialVersion is the initial schema version for every server.
 	// It's exported for testing.
-	InitialVersion           = "0"
-	putKeyNoRetry            = 1
-	keyOpDefaultRetryCnt     = 3
-	putKeyRetryUnlimited     = math.MaxInt64
-	keyOpDefaultTimeout      = 2 * time.Second
-	keyOpRetryInterval       = 30 * time.Millisecond
-	checkVersInterval        = 20 * time.Millisecond
-	waitTimeWhenErrorOccured = 1 * time.Second
+	InitialVersion       = "0"
+	putKeyNoRetry        = 1
+	keyOpDefaultRetryCnt = 3
+	putKeyRetryUnlimited = math.MaxInt64
+	keyOpDefaultTimeout  = 2 * time.Second
+	keyOpRetryInterval   = 30 * time.Millisecond
+	checkVersInterval    = 20 * time.Millisecond
 )
 
 var (
@@ -55,6 +54,8 @@ var (
 	// SyncerSessionTTL is the etcd session's TTL in seconds.
 	// and it's an exported variable for testing.
 	SyncerSessionTTL = 10 * 60
+	// WaitTimeWhenErrorOccured is waiting interval when processing DDL jobs encounter errors.
+	WaitTimeWhenErrorOccured = 1 * time.Second
 )
 
 // SchemaSyncer is used to synchronize schema version between the DDL worker leader and followers through etcd.