Thank you for working on TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.

## What have you changed? (mandatory)

Reduce test time when running tests with `race`
- reduce the value of `WaitTimeWhenErrorOccured`
- use a different database name than other tests, like the test of `TestCreateTableWithLike` and `TestUpdateMultipleTable`

Move the `testDropColumn2` out of `TestColumn`. It's easier to local the problem when the test encounters an error.


## What are the type of the changes (mandatory)?
Improve test

## Does this PR affect documentation (docs/docs-cn) update? (optional)

No.

## Benchmark result if necessary (optional)
TestCreateTableWithLike 
Before:
PASS: ddl_db_test.go:1464: testDBSuite.TestCreateTableWithLike	19.081s
PASS: ddl_db_test.go:1466: testDBSuite.TestCreateTableWithLike	17.106s
After:
PASS: ddl_db_test.go:1467: testDBSuite.TestCreateTableWithLike	10.115s
PASS: ddl_db_test.go:1466: testDBSuite.TestCreateTableWithLike	8.575s

TestUpdateMultipleTable
Before:
PASS: ddl_db_test.go:1386: testDBSuite.TestUpdateMultipleTable	8.392s
PASS: ddl_db_test.go:1386: testDBSuite.TestUpdateMultipleTable	7.681s
After:
PASS: ddl_db_test.go:1390: testDBSuite.TestUpdateMultipleTable	4.524s
PASS: ddl_db_test.go:1389: testDBSuite.TestUpdateMultipleTable	4.907s

testDDLSuite
Before:
ok  	github.com/pingcap/tidb/ddl	30.046s
After:
ok  	github.com/pingcap/tidb/ddl	5.334s


All tests in the package of ddl
Before:
ok  	github.com/pingcap/tidb/ddl	212.934s
ok  	github.com/pingcap/tidb/ddl	217.840s
After:
ok  	github.com/pingcap/tidb/ddl	184.299s
ok  	github.com/pingcap/tidb/ddl	180.139s