@shenli 2018-06-12
It is wired that changing dbname could speed up tests.