@zimulala 2018-06-12
/run-all-tests